+++ lib/src/main/java/com/diffplug/spotless/kotlin/KtfmtStep.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2024 DiffPlug
+ * Copyright 2016-2025 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ package com.diffplug.spotless.kotlin;
 import static com.diffplug.spotless.kotlin.KtfmtStep.Style.DEFAULT;
 import static com.diffplug.spotless.kotlin.KtfmtStep.Style.DROPBOX;
 import static com.diffplug.spotless.kotlin.KtfmtStep.Style.META;
+import static com.diffplug.spotless.kotlin.KtfmtStep.TrailingCommaManagementStrategy.ONLY_ADD;
 
 import java.io.Serializable;
 import java.lang.reflect.Constructor;
@@ -39,7 +40,7 @@ import com.diffplug.spotless.ThrowingEx;
  */
 public class KtfmtStep implements Serializable {
 	private static final long serialVersionUID = 1L;
-	private static final String DEFAULT_VERSION = "0.53";
+	private static final String DEFAULT_VERSION = "0.58";
 	private static final String NAME = "ktfmt";
 	private static final String MAVEN_COORDINATE = "com.facebook:ktfmt:";
 
@@ -47,11 +48,11 @@ public class KtfmtStep implements Serializable {
 	/**
 	 * Option that allows to apply formatting options to perform a 4-space block and continuation indent.
 	 */
-	@Nullable
-	private final Style style;
-	@Nullable
-	private final KtfmtFormattingOptions options;
-	/** The jar that contains the formatter. */
+	@Nullable private final Style style;
+	@Nullable private final KtfmtFormattingOptions options;
+	/**
+	 * The jar that contains the formatter.
+	 */
 	private final JarState.Promised jarState;
 
 	private KtfmtStep(String version,
@@ -103,31 +104,55 @@ public class KtfmtStep implements Serializable {
 			return since;
 		}
 
-		/** Last version (inclusive) that supports this style */
-		@Nullable
-		String getUntil() {
+		/**
+		 * Last version (inclusive) that supports this style
+		 */
+		@Nullable String getUntil() {
 			return until;
 		}
+	}
+
+	public enum TrailingCommaManagementStrategy {
+		/**
+		 * Do not manage trailing commas at all, only format what is already present.
+		 */
+		NONE,
+		/**
+		 * <p>
+		 * Only add trailing commas when necessary, but do not remove them.
+		 * </p>
+		 * <p>
+		 * Lists that cannot fit on one line will have trailing commas inserted.
+		 * Trailing commas can to be used to "hint" ktfmt that the list should be broken to multiple lines
+		 * </p>
+		 */
+		ONLY_ADD,
+		/**
+		 * <p>
+		 * Fully manage trailing commas, adding and removing them where necessary.
+		 * </p>
+		 * <p>
+		 * Lists that cannot fit on one line will have trailing commas inserted.
+		 * Lists that span multiple lines will have them removed. Manually inserted trailing commas
+		 * cannot be used as a hint to force breaking lists to multiple lines.
+		 * </p>
+		 */
+		COMPLETE,
 	}
 
 	public static class KtfmtFormattingOptions implements Serializable {
 
 		private static final long serialVersionUID = 1L;
 
-		@Nullable
-		private Integer maxWidth = null;
+		@Nullable private Integer maxWidth = null;
 
-		@Nullable
-		private Integer blockIndent = null;
+		@Nullable private Integer blockIndent = null;
 
-		@Nullable
-		private Integer continuationIndent = null;
+		@Nullable private Integer continuationIndent = null;
 
-		@Nullable
-		private Boolean removeUnusedImports = null;
+		@Nullable private Boolean removeUnusedImports = null;
 
-		@Nullable
-		private Boolean manageTrailingCommas = null;
+		@Nullable private TrailingCommaManagementStrategy trailingCommaManagementStrategy;
 
 		public KtfmtFormattingOptions() {}
 
@@ -136,12 +161,12 @@ public class KtfmtStep implements Serializable {
 				@Nullable Integer blockIndent,
 				@Nullable Integer continuationIndent,
 				@Nullable Boolean removeUnusedImports,
-				@Nullable Boolean manageTrailingCommas) {
+				@Nullable TrailingCommaManagementStrategy trailingCommaManagementStrategy) {
 			this.maxWidth = maxWidth;
 			this.blockIndent = blockIndent;
 			this.continuationIndent = continuationIndent;
 			this.removeUnusedImports = removeUnusedImports;
-			this.manageTrailingCommas = manageTrailingCommas;
+			this.trailingCommaManagementStrategy = trailingCommaManagementStrategy;
 		}
 
 		public void setMaxWidth(int maxWidth) {
@@ -160,22 +185,28 @@ public class KtfmtStep implements Serializable {
 			this.removeUnusedImports = removeUnusedImports;
 		}
 
-		public void setManageTrailingCommas(boolean manageTrailingCommas) {
-			this.manageTrailingCommas = manageTrailingCommas;
+		public void setTrailingCommaManagementStrategy(TrailingCommaManagementStrategy trailingCommaManagementStrategy) {
+			this.trailingCommaManagementStrategy = trailingCommaManagementStrategy;
 		}
 	}
 
-	/** Creates a step which formats everything - code, import order, and unused imports. */
+	/**
+	 * Creates a step which formats everything - code, import order, and unused imports.
+	 */
 	public static FormatterStep create(Provisioner provisioner) {
 		return create(defaultVersion(), provisioner);
 	}
 
-	/** Creates a step which formats everything - code, import order, and unused imports. */
+	/**
+	 * Creates a step which formats everything - code, import order, and unused imports.
+	 */
 	public static FormatterStep create(String version, Provisioner provisioner) {
 		return create(version, provisioner, null, null);
 	}
 
-	/** Creates a step which formats everything - code, import order, and unused imports. */
+	/**
+	 * Creates a step which formats everything - code, import order, and unused imports.
+	 */
 	public static FormatterStep create(String version, Provisioner provisioner, @Nullable Style style, @Nullable KtfmtFormattingOptions options) {
 		Objects.requireNonNull(version, "version");
 		Objects.requireNonNull(provisioner, "provisioner");
@@ -196,10 +227,8 @@ public class KtfmtStep implements Serializable {
 	private static final class State implements Serializable {
 		private static final long serialVersionUID = 1L;
 		private final String version;
-		@Nullable
-		private final Style style;
-		@Nullable
-		private final KtfmtFormattingOptions options;
+		@Nullable private final Style style;
+		@Nullable private final KtfmtFormattingOptions options;
 		private final JarState jarState;
 
 		State(String version,
@@ -224,6 +253,7 @@ public class KtfmtStep implements Serializable {
 			final Class<?> formatterFuncClass = classLoader.loadClass("com.diffplug.spotless.glue.ktfmt.KtfmtFormatterFunc");
 			final Class<?> ktfmtStyleClass = classLoader.loadClass("com.diffplug.spotless.glue.ktfmt.KtfmtStyle");
 			final Class<?> ktfmtFormattingOptionsClass = classLoader.loadClass("com.diffplug.spotless.glue.ktfmt.KtfmtFormattingOptions");
+			final Class<?> ktfmtTrailingCommaManagmentStrategyClass = classLoader.loadClass("com.diffplug.spotless.glue.ktfmt.KtfmtTrailingCommaManagementStrategy");
 
 			if (style == null && options == null) {
 				final Constructor<?> constructor = formatterFuncClass.getConstructor();
@@ -237,9 +267,13 @@ public class KtfmtStep implements Serializable {
 			}
 
 			final Constructor<?> optionsConstructor = ktfmtFormattingOptionsClass.getConstructor(
-					Integer.class, Integer.class, Integer.class, Boolean.class, Boolean.class);
+					Integer.class, Integer.class, Integer.class, Boolean.class, ktfmtTrailingCommaManagmentStrategyClass);
+
+			final Object ktfmtTrailingCommaManagementStrategy = options.trailingCommaManagementStrategy == null
+					? null
+					: Enum.valueOf((Class<? extends Enum>) ktfmtTrailingCommaManagmentStrategyClass, options.trailingCommaManagementStrategy.name());
 			final Object ktfmtFormattingOptions = optionsConstructor.newInstance(
-					options.maxWidth, options.blockIndent, options.continuationIndent, options.removeUnusedImports, options.manageTrailingCommas);
+					options.maxWidth, options.blockIndent, options.continuationIndent, options.removeUnusedImports, ktfmtTrailingCommaManagementStrategy);
 			if (style == null) {
 				final Constructor<?> constructor = formatterFuncClass.getConstructor(ktfmtFormattingOptionsClass);
 				return (FormatterFunc) constructor.newInstance(ktfmtFormattingOptions);
@@ -260,6 +294,12 @@ public class KtfmtStep implements Serializable {
 			if (BadSemver.version(version) < BadSemver.version(0, 17)) {
 				if (options != null && options.removeUnusedImports != null) {
 					throw new IllegalStateException("Ktfmt formatting option `removeUnusedImports` supported for version 0.17 and later");
+				}
+			}
+
+			if (BadSemver.version(version) < BadSemver.version(0, 57)) {
+				if (options != null && options.trailingCommaManagementStrategy == ONLY_ADD) {
+					throw new IllegalStateException("Value ONLY_ADD for Ktfmt formatting option `trailingCommaManagementStrategy` supported for version 0.57 and later");
 				}
 			}
 		}
@@ -365,7 +405,7 @@ public class KtfmtStep implements Serializable {
 							/* continuationIndent = */ Optional.ofNullable(options.continuationIndent).orElse((Integer) formattingOptionsClass.getMethod("getContinuationIndent").invoke(formattingOptions)),
 							/* removeUnusedImports = */ Optional.ofNullable(options.removeUnusedImports).orElse((Boolean) formattingOptionsClass.getMethod("getRemoveUnusedImports").invoke(formattingOptions)),
 							/* debuggingPrintOpsAfterFormatting = */ (Boolean) formattingOptionsClass.getMethod("getDebuggingPrintOpsAfterFormatting").invoke(formattingOptions));
-				} else {
+				} else if (BadSemver.version(version) < BadSemver.version(0, 57)) {
 					Class<?> styleClass = classLoader.loadClass(formattingOptionsClass.getName() + "$Style");
 					formattingOptions = formattingOptions.getClass().getConstructor(styleClass, int.class, int.class, int.class, boolean.class, boolean.class, boolean.class).newInstance(
 							/* style = */ formattingOptionsClass.getMethod("getStyle").invoke(formattingOptions),
@@ -374,7 +414,17 @@ public class KtfmtStep implements Serializable {
 							/* continuationIndent = */ Optional.ofNullable(options.continuationIndent).orElse((Integer) formattingOptionsClass.getMethod("getContinuationIndent").invoke(formattingOptions)),
 							/* removeUnusedImports = */ Optional.ofNullable(options.removeUnusedImports).orElse((Boolean) formattingOptionsClass.getMethod("getRemoveUnusedImports").invoke(formattingOptions)),
 							/* debuggingPrintOpsAfterFormatting = */ (Boolean) formattingOptionsClass.getMethod("getDebuggingPrintOpsAfterFormatting").invoke(formattingOptions),
-							/* manageTrailingCommas */ Optional.ofNullable(options.manageTrailingCommas).orElse((Boolean) formattingOptionsClass.getMethod("getManageTrailingCommas").invoke(formattingOptions)));
+							/* manageTrailingCommas = */ Optional.ofNullable(getManageTrailingCommasFrom(options.trailingCommaManagementStrategy)).orElse((Boolean) formattingOptionsClass.getMethod("getManageTrailingCommas").invoke(formattingOptions)));
+				} else {
+					Class<?> styleClass = classLoader.loadClass(formattingOptionsClass.getName() + "$Style");
+					formattingOptions = formattingOptions.getClass().getConstructor(styleClass, int.class, int.class, int.class, boolean.class, boolean.class, TrailingCommaManagementStrategy.class).newInstance(
+							/* style = */ formattingOptionsClass.getMethod("getStyle").invoke(formattingOptions),
+							/* maxWidth = */ Optional.ofNullable(options.maxWidth).orElse((Integer) formattingOptionsClass.getMethod("getMaxWidth").invoke(formattingOptions)),
+							/* blockIndent = */ Optional.ofNullable(options.blockIndent).orElse((Integer) formattingOptionsClass.getMethod("getBlockIndent").invoke(formattingOptions)),
+							/* continuationIndent = */ Optional.ofNullable(options.continuationIndent).orElse((Integer) formattingOptionsClass.getMethod("getContinuationIndent").invoke(formattingOptions)),
+							/* removeUnusedImports = */ Optional.ofNullable(options.removeUnusedImports).orElse((Boolean) formattingOptionsClass.getMethod("getRemoveUnusedImports").invoke(formattingOptions)),
+							/* debuggingPrintOpsAfterFormatting = */ (Boolean) formattingOptionsClass.getMethod("getDebuggingPrintOpsAfterFormatting").invoke(formattingOptions),
+							/* trailingCommaManagementStrategy */ Optional.ofNullable(options.trailingCommaManagementStrategy).orElse((TrailingCommaManagementStrategy) formattingOptionsClass.getMethod("getTrailingCommaManagementStrategy").invoke(formattingOptions)));
 				}
 			}
 
@@ -421,6 +471,19 @@ public class KtfmtStep implements Serializable {
 				formattingOptionsClazz = classLoader.loadClass(PACKAGE + ".FormattingOptions");
 			}
 			return formattingOptionsClazz;
+		}
+
+		private @Nullable Boolean getManageTrailingCommasFrom(
+			@Nullable TrailingCommaManagementStrategy trailingCommaManagementStrategy
+		) {
+			if (trailingCommaManagementStrategy == null) {
+				return null;
+			}
+
+			return switch (trailingCommaManagementStrategy) {
+				case NONE, ONLY_ADD -> false;
+				case COMPLETE -> true;
+			};
 		}
 	}
 }